Extract redmine tasks into own testcase class, add composer script for e2e tests
